# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Template API for Golang project",
        "title": "Base APP Go",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/roles/{role_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches the details of a single role by its ID. Requires Bearer token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Get Role Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get role detail",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_role_presenter.ReadRoleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid Role ID)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found (Role ID not found)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Failed get role detail",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing role by its ID. Requires Bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Update Role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role data to update",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_role_presenter.UpdateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update role",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid ID, Invalid JSON, or Validation Error)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Role ID not found",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Failed update role",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a role by its ID. Requires Bearer token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Delete Role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete role",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid Role ID)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (Missing Auth)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Role ID not found",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Failed delete role",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            }
        },
...
        "/users/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches the details of a single user by their ID. Requires Bearer token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get user detail",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_user_presenter.ReadUserDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid User ID)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found (User ID not found)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a user's profile details by their ID. Requires Bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User data to update",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_user_presenter.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update user",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid ID, Invalid JSON, or Validation Error)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "User ID not found",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a user by their ID. Requires Bearer token. Users cannot delete themselves.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete user",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid User ID)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "403": {
                        "description": "Forbidden (Attempt to delete own account)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found (User ID not found)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "be-dashboard-nba_internal_presentation_response.DataPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_response.ErrorPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "be-dashboard-nba_internal_presentation_role_presenter.ReadRoleResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "be-dashboard-nba_internal_presentation_role_presenter.UpdateRoleRequest": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "be-dashboard-nba_internal_presentation_user_presenter.ReadUserDetailResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                }
            }
        },
...
        "be-dashboard-nba_internal_presentation_user_presenter.UpdateUserRequest": {
            "type": "object",
            "required": [
                "active",
                "email",
                "full_name",
                "name",
                "role_id"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {your token}\" into the field below.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}